test(runner): split sandbox run tests by behavior
